2019/04/16 update: addr2line64 added, Julie's Mac stack trace fix, autograder warnings, etc.

+++ StanfordCPPLib/util/memory.cpp
@@ -5,6 +5,8 @@
  * allocation.
  *
  * @author Marty Stepp
+ * @version 2019/04/16
+ * - bug fixes for WIN64
  * @version 2018/12/04
  * - initial version
  */
@@ -19,40 +21,40 @@ namespace memory {
 
 void computeMemoryDistances(
         void* const p,
-        unsigned long int& stackDist,
-        unsigned long int& heapDist,
-        unsigned long int& staticDist) {
+        uintptr_t& stackDist,
+        uintptr_t& heapDist,
+        uintptr_t& staticDist) {
     static int si = 0;
     int i = 0;
     int* pi = new int(0);
 
     // don't use abs/subtraction here because it can over/underflow with very large integers
     stackDist = p > &i ?
-            reinterpret_cast<unsigned long int>(p) - reinterpret_cast<unsigned long int>(&i) :
-            reinterpret_cast<unsigned long int>(&i) - reinterpret_cast<unsigned long int>(p);
+            reinterpret_cast<uintptr_t>(p) - reinterpret_cast<uintptr_t>(&i) :
+            reinterpret_cast<uintptr_t>(&i) - reinterpret_cast<uintptr_t>(p);
     heapDist = p > pi ?
-            reinterpret_cast<unsigned long int>(p) - reinterpret_cast<unsigned long int>(pi) :
-            reinterpret_cast<unsigned long int>(pi) - reinterpret_cast<unsigned long int>(p);
+            reinterpret_cast<uintptr_t>(p) - reinterpret_cast<uintptr_t>(pi) :
+            reinterpret_cast<uintptr_t>(pi) - reinterpret_cast<uintptr_t>(p);
     staticDist = p > &si ?
-            reinterpret_cast<unsigned long int>(p) - reinterpret_cast<unsigned long int>(&si) :
-            reinterpret_cast<unsigned long int>(&si) - reinterpret_cast<unsigned long int>(p);
+            reinterpret_cast<uintptr_t>(p) - reinterpret_cast<uintptr_t>(&si) :
+            reinterpret_cast<uintptr_t>(&si) - reinterpret_cast<uintptr_t>(p);
     delete pi;
 }
 
 bool isOnHeap(void* const p) {
-    unsigned long int stackDist, heapDist, staticDist;
+    uintptr_t stackDist, heapDist, staticDist;
     computeMemoryDistances(p, stackDist, heapDist, staticDist);
     return heapDist < stackDist && heapDist < staticDist;
 }
 
 bool isOnStack(void* const p) {
-    unsigned long int stackDist, heapDist, staticDist;
+    uintptr_t stackDist, heapDist, staticDist;
     computeMemoryDistances(p, stackDist, heapDist, staticDist);
     return stackDist < heapDist && stackDist < staticDist;
 }
 
 bool isOnStatic(void* const p) {
-    unsigned long int stackDist, heapDist, staticDist;
+    uintptr_t stackDist, heapDist, staticDist;
     computeMemoryDistances(p, stackDist, heapDist, staticDist);
     return staticDist < stackDist && staticDist < heapDist;
 }
